Deleting now feature complete
An admin user can now delete aany other user. A error will show if they try to delete their own account, stating that they cannot do that. As well as that, the auto increment ids in the database for each account are reset to not have any gaps when an account is deleted.